Add test for running on worker threads
